+++ application/controllers/tela.php
@@ -13,6 +13,11 @@ class Tela extends CI_Controller {
 	
 	public function index()
 	{
+		$this->load->model('perfil_model');
+		if (!$this->perfil_model->verifica_acesso($this->registro,__METHOD__))
+		{
+			header('location:../forbidden');exit;
+		}
 		$head = array();
 		$head['title'] = 'Tela';
 		$this->load->view('header', $head);
@@ -28,6 +33,11 @@ class Tela extends CI_Controller {
 
 	public function edit($id = null)
 	{
+		$this->load->model('perfil_model');
+		if (!$this->perfil_model->verifica_acesso($this->registro,__METHOD__))
+		{
+			header('location:../forbidden');exit;
+		}
 		$head = array();
 		$head['title'] = 'Tela';
 		$this->load->view('header', $head);
@@ -47,6 +57,11 @@ class Tela extends CI_Controller {
 
 	public function delete()
 	{
+		$this->load->model('perfil_model');
+		if (!$this->perfil_model->verifica_acesso($this->registro,__METHOD__))
+		{
+			header('location:../forbidden');exit;
+		}
 		$this->load->model('tela_model');
 		$id = $this->input->post('id');
 		$msg = $this->tela_model->delete($id);
@@ -76,6 +91,11 @@ TXT;
 
 	public function save()
 	{
+		$this->load->model('perfil_model');
+		if (!$this->perfil_model->verifica_acesso($this->registro,__METHOD__))
+		{
+			header('location:../forbidden');exit;
+		}
 		$this->load->model('tela_model');
 
 		$head = array();
